% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_ASI_engagements.R
\name{add_ASI_engagements}
\alias{add_ASI_engagements}
\title{Add engagements to ASI file}
\usage{
add_ASI_engagements(asiText, scenario, drawSensorTracks = FALSE)
}
\arguments{
\item{asiText}{output from create_ASI_platforms (can be piped into this function)}

\item{scenario}{must have targetTruth and/or ownShipTruth as well as engagementData:}

\item{drawSensorTracks}{(default=FALSE) needed for proper PlatformID numbering}
}
\value{
text
}
\description{
Add engagements to ASI file
}
