% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_ASI_header.R
\name{create_ASI_header}
\alias{create_ASI_header}
\title{Create the header for an ASI file}
\usage{
create_ASI_header(prefFile = NA)
}
\arguments{
\item{prefFile}{optional. Adds a preference file title to the header  (default=NA)}
}
\value{
header text
}
\description{
Create the header for an ASI file
}
