% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_ASI_platforms.R
\name{create_ASI_platforms}
\alias{create_ASI_platforms}
\title{Create an ASI platform}
\usage{
create_ASI_platforms(header, scenario, persistPosition = FALSE,
  drawSensorTracks = FALSE)
}
\arguments{
\item{header}{output from create_ASI_header (can be piped into this function)}

\item{scenario}{must have targetTruth and/or ownShipTruth as well as platformInfo:

platformData (targetTruth and/or ownShipTruth) position data on all the platforms. This is the Nautilus truth format. Requires the following columns:
\itemize{
 \item{time: (double) time of measurement. We currently recommend POSIX}
 \item{lon: (double) longitude of target at time of measurement}
 \item{lat: (double) latitude of target at time of measurement}
 \item{alt: (double) altitude of target at time of measurement}
 \item{truthID: (factor) name or identifier for target. We recommend letters or names}
 \item{heading: (double) sensor system heading in degrees azimuth}
 }

platformInfo display info for all platforms. Requires the following columns:
\itemize{
 \item{truthID: (factor) name or identifier for target}
 \item{platformIcon: (string) name of platform icon (must choose from valid SIMDIS options)}
 \item{platformType: (string) type of platform (e.g., ship) (must choose from valid SIMDIS options)}
 \item{trackColor: (string) color of the track (must be a valid SIMDIS choice)}
 }

 @param persistPosition (boolean) if true, will persist all platforms until this time (and gray out the title at their last position) (default=false)}

\item{drawSensorTracks}{(default=FALSE) if the scenario includes sensorData, the tracks will be drawn in the asi file (it's okay if drawSensorTracks=TRUE but no sensorData exists in the scenario)}
}
\value{
text
}
\description{
Gives text for platform instantiation
}
