% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_scenario_to_ASI.R
\name{export_scenario_to_ASI}
\alias{export_scenario_to_ASI}
\title{Writes your scenario as an ASI file}
\usage{
export_scenario_to_ASI(scenario, outFileName, drawEngagements = TRUE,
  drawSensorTracks = FALSE, persistPosition = FALSE, prefFile = NA)
}
\arguments{
\item{scenario}{must have targetTruth and/or ownShipTruth as well as platformInfo}

\item{outFileName}{output file name (including path if desired) for the ASI. Optional to include '.asi' in the output}

\item{drawEngagements}{(default=TRUE) if the scenario includes engagementData, the engagements will be drawn in the asi file (it's okay if drawEngagements=TRUE but no engagementData exists in the scenario)}

\item{drawSensorTracks}{(default=FALSE) if the scenario includes sensorData, the tracks will be drawn in the asi file (it's okay if drawSensorTracks=TRUE but no sensorData exists in the scenario)}

\item{persistPosition}{(boolean) if true, will persist all platforms until the end of the file (and gray out the title at their last position) (default=false)}

\item{prefFile}{(default=NA) specify name of preference file (must be in same folder as the ASI output when you open it in simdis)}
}
\description{

}
