% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_ASI_position.R
\name{format_ASI_position}
\alias{format_ASI_position}
\title{Format scenarioMaker data into SIMDIS position data}
\usage{
format_ASI_position(platformData, timeOverride = 0)
}
\arguments{
\item{platformData}{position data on all the platforms. This is the scenarioMaker truth format. Requires the following columns:
\itemize{
 \item{time: (double) time of measurement. We currently recommend POSIX}
 \item{lon: (double) longitude of target at time of measurement}
 \item{lat: (double) latitude of target at time of measurement}
 \item{alt: (double) altitude of target at time of measurement}
 \item{truthID: (factor) name or identifier for target. We recommend letters or names}
 \item{heading: (double) sensor system heading in degrees azimuth}
 }}

\item{timeOverride}{if non-zero, will overwrite the time in platformData (default=0) (used for the persistPosition option in create_ASI_platforms)}
}
\value{
platform data for SIMDIS
}
\description{
Format scenarioMaker data into SIMDIS position data
}
