% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_add_engagements.R
\name{plot_add_engagements}
\alias{plot_add_engagements}
\title{Plot SUW combat money chart}
\usage{
plot_add_engagements(scenario, useDefaultColors = TRUE)
}
\arguments{
\item{scenario}{must contain the following items:}

\item{useDefaultColors}{(default=FALSE) if you DO have platformInfo filled in, but want to ignore the colors and use defaults, set this to true
targetTruth - (scenarioMaker format)
\itemize{
 \item{time: (double) time of measurement. We currently recommend POSIX}
 \item{lon: (double) longitude of target at time of measurement}
 \item{lat: (double) latitude of target at time of measurement}
 \item{alt: (double) altitude of target at time of measurement}
 \item{truthID: (factor) name or identifier for target. We recommend letters or names}
 \item{heading: (double) target heading in degrees azimuth}
 }
 ownShipTruth - (scenarioMaker format)
\itemize{
 \item{time: (double) time of measurement. We currently recommend POSIX}
 \item{lon: (double) longitude of target at time of measurement}
 \item{lat: (double) latitude of target at time of measurement}
 \item{alt: (double) altitude of target at time of measurement}
 \item{truthID: (factor) name or identifier for target. We recommend letters or names}
 \item{heading: (double) target heading in degrees azimuth}
 }
engagementData - who shot whom, with what, when, what color, and did it kill them?
\itemize{
 \item{time: (double) time of measurement. We currently recommend POSIX}
 \item{source: (factor) the truthID of the shooter}
 \item{target: (factor) the truthID of the thing being shot}
 \item{weapon: (factor) name of the weapon}
 \item{color: (string) color of the weapon (this color will be used in ggplot, plotly, and exported to SIMDIS) (this function doesn't yet respect color)}
 \item{kill: (double) if the engagement killed the target, kill = 1, otherwise kill = 0}
}}
}
\value{
plotly object
}
\description{
This figure plots target range as a function of time
}
\examples{
plot_add_engagements(scenarioMaker::example1_scenario)
}
