% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ASI.R
\name{read_ASI}
\alias{read_ASI}
\title{Read platform data (truth data) from an ASI file}
\usage{
read_ASI(ASIFilePath, timeFormat = "\%j \%Y \%H:\%M:\%OS", timeZone = NULL)
}
\arguments{
\item{ASIFilePath}{path to ASI file}

\item{timeFormat}{(default="\%j \%Y \%H:\%M:\%OS") the format the time information is written in - if it's just seconds, use "\%OS\%"}

\item{timeZone}{(default=NULL) specify string of time zone. If no argument supplied, will actually default to UTC, but will warn the user}
}
\value{
truth data dataframe
}
\description{
Read platform data (truth data) from an ASI file. Only reads in lat, lon, alt, and heading (no speed, velocity, etc.).
Only works on very SIMDIS files written in an LLA coordinate system. If you export your SIMDIS file (from the File menu), it will save it in a way that this can read
}
\examples{
\dontrun{read_ASI("C:/myFile.asi",timeZone="US/Pacific")}

}
