% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sandTable_interface.R
\name{sandTable_interface}
\alias{sandTable_interface}
\title{sandTable interface function (shiny)}
\usage{
sandTable_interface()
}
\value{
none
}
\description{
Function to create shiny interface
}
\keyword{internal}
